\name{flexmixNL-internal}
\alias{FLXgetModelmatrix,FLXMRnlm-method}
\alias{FLXdeterminePostunscaled,FLXMRnlm-method}
\alias{FLXMRnlm-class}
\alias{FLXmstep,FLXMRnlm-method}       
\alias{gnm.wfit}
\alias{nls.wfit}

\title{Internal flexmixNL Functions}
\description{Internal \code{\link{flexmixNL}} functions, methods and classes.}
\details{These are not to be called by the user.}
\keyword{internal}